backend/services/ranking-service/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
scipy==1.11.2